[refactor] Input 및 Modal의 X 버튼 아이콘 일원화 (#206)
* refactor: ModalHeader 및 Input에 있는 X 버튼 통합

* design: Input에 있는 X 버튼 색상 밑줄 색상과 통일

* chore: CloseIcon import 절대경로로 수정

절대경로를 사용하면 파일 위치가 변경되어도 import를 수정할 필요가 없어짐

* chore: interface -> type으로 수정

* fix: Layout.stories에 있는 CloseIcon import 오류 수정

* refactor: useTheme을 사용하여 컴포넌트 내부 테마 적용
