+++ frontend/src/components/@common/Modal/ModalHeader/ModalHeader.tsx
@@ -1,5 +1,6 @@
-import CloseIcon from '@/assets/close.svg';
+import CloseIcon from '@/components/@common/CloseIcon/CloseIcon';
 import Headline3 from '@/components/@common/Headline3/Headline3';
+import { useTheme } from '@emotion/react';
 import * as S from './ModalHeader.styled';
 
 type Props = {
@@ -9,12 +10,14 @@ type Props = {
 };
 
 const ModalHeader = ({ title, onClose, showCloseButton = true }: Props) => {
+  const theme = useTheme();
+
   return (
     <S.Container>
       <Headline3>{title}</Headline3>
       {showCloseButton && (
         <S.CloseButton onClick={onClose}>
-          <S.CloseIcon src={CloseIcon} alt="close-icon" />
+          <CloseIcon stroke={theme.color.gray[600]} />
         </S.CloseButton>
       )}
     </S.Container>
