+++ frontend/src/styles/emotion.d.ts
@@ -2,8 +2,8 @@ import '@emotion/react';
 import { theme } from './theme';
 
 declare module '@emotion/react' {
-  export type Theme = {
+  export interface Theme {
     color: typeof theme.color;
     typography: typeof theme.typography;
-  };
+  }
 }
